+++ src/model/Application.zig
@@ -3,6 +3,8 @@ const zigcord = @import("../root.zig");
 const jconfig = @import("../root.zig").jconfig;
 const model = zigcord.model;
 
+const Application = @This();
+
 id: model.Snowflake,
 name: []const u8,
 icon: ?[]const u8,
@@ -25,10 +27,14 @@ flags: jconfig.Omittable(Flags) = .omit,
 approximate_guild_count: jconfig.Omittable(i64) = .omit,
 approximate_user_install_count: jconfig.Omittable(i64) = .omit,
 redirect_uris: jconfig.Omittable([]const []const u8) = .omit,
-interactions_endpoint_url: jconfig.Omittable([]const u8) = .omit,
-role_connections_verification_url: jconfig.Omittable([]const u8) = .omit,
+interactions_endpoint_url: jconfig.Omittable(?[]const u8) = .omit,
+role_connections_verification_url: jconfig.Omittable(?[]const u8) = .omit,
+event_webhooks_url: jconfig.Omittable(?[]const u8) = .omit,
+event_webhooks_status: jconfig.Omittable(ApplicationEventWebhookStatus) = .omit,
+event_webhooks_types: jconfig.Omittable([]const u8) = .omit,
 tags: jconfig.Omittable([]const []const u8) = .omit,
 install_params: jconfig.Omittable(InstallParams) = .omit,
+integration_types_config: jconfig.Omittable(IntegrationTypeConfigurationDict) = .omit,
 custom_install_url: jconfig.Omittable([]const u8) = .omit,
 
 pub const jsonStringify = jconfig.stringifyWithOmit;
@@ -44,8 +50,11 @@ pub const Team = struct {
 pub const TeamMember = struct {
     membership_state: State,
     team_id: model.Snowflake,
-    user: model.User,
+    user: jconfig.Partial(model.User),
     role: []const u8,
+    permissions: jconfig.Omittable([]const []const u8) = .omit,
+
+    pub const jsonStringify = jconfig.stringifyWithOmit;
 
     pub const State = enum(u8) {
         invited = 1,
@@ -94,6 +103,14 @@ pub const Flags = packed struct(u64) {
     }
 };
 
+pub const ApplicationEventWebhookStatus = enum(u8) {
+    disabled = 1,
+    enabled = 2,
+    disabled_by_discord = 3,
+
+    pub const jsonStringify = jconfig.stringifyEnumAsInt;
+};
+
 pub const InstallParams = struct {
     scopes: []const []const u8,
     permissions: []const u8,
@@ -105,3 +122,77 @@ pub const IntegrationType = enum {
 
     pub const jsonStringify = jconfig.stringifyEnumAsInt;
 };
+
+pub const IntegrationTypeConfigurationDict = struct {
+    guild_install: jconfig.Omittable(IntegrationTypeConfiguration) = .omit,
+    user_install: jconfig.Omittable(IntegrationTypeConfiguration) = .omit,
+
+    pub fn jsonParse(allocator: std.mem.Allocator, source: anytype, options: std.json.ParseOptions) !IntegrationTypeConfigurationDict {
+        const obj_token = try source.nextAlloc(allocator, options.allocate orelse .alloc_if_needed);
+        switch (obj_token) {
+            .object_begin => {},
+            else => return error.UnexpectedToken,
+        }
+
+        var result: IntegrationTypeConfigurationDict = .{};
+
+        while (true) {
+            const field_name_token = try source.nextAlloc(allocator, options.allocate orelse .alloc_if_needed);
+            const field_name = switch (field_name_token) {
+                .string => |str| str,
+                .allocated_string => |str| str,
+                .object_end => {
+                    return result;
+                },
+                else => return error.UnexpectedToken,
+            };
+            const field_value: jconfig.Omittable(IntegrationTypeConfiguration) =
+                .initSome(try std.json.innerParse(IntegrationTypeConfiguration, allocator, source, options));
+
+            if (std.mem.eql(u8, field_name, "0")) {
+                result.guild_install = field_value;
+            } else if (std.mem.eql(u8, field_name, "1")) {
+                result.user_install = field_value;
+            }
+        }
+    }
+
+    pub fn jsonParseByValue(allocator: std.mem.Allocator, source: std.json.Value, options: std.json.ParseOptions) !IntegrationTypeConfigurationDict {
+        var obj = switch (source) {
+            .object => |obj| obj,
+            else => return error.UnexpectedToken,
+        };
+
+        var result: IntegrationTypeConfigurationDict = .{};
+        if (obj.get("0")) |guild_install_value| {
+            result.guild_install = .initSome(try std.json.innerParseFromValue(IntegrationTypeConfiguration, allocator, guild_install_value, options));
+        }
+        if (obj.get("1")) |user_install_value| {
+            result.user_install = .initSome(try std.json.innerParseFromValue(IntegrationTypeConfiguration, allocator, user_install_value, options));
+        }
+        return result;
+    }
+
+    pub fn jsonStringify(self: IntegrationTypeConfigurationDict, jw: *std.json.Stringify) !void {
+        try jw.beginObject();
+        if (self.guild_install.asSome()) |config| {
+            try jw.objectField("0");
+            try jw.write(config);
+        }
+        if (self.user_install.asSome()) |config| {
+            try jw.objectField("1");
+            try jw.write(config);
+        }
+        try jw.endObject();
+    }
+};
+
+pub const IntegrationTypeConfiguration = struct {
+    oauth2_install_params: InstallParams,
+};
+
+// from https://discord.com/developers/docs/resources/application#application-object-example-application-object
+test "redacted enkafang" {
+    const input = @embedFile("./test/application.test.json");
+    try jconfig.testing.expectParsedSuccessfully(Application, std.testing.allocator, input, .{ .ignore_unknown_fields = true });
+}